+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.1.5] - 2019-08-14
+
+### Misc
+
+-   Change Logo
+-   Change loading icon in terminus
+-   Prefix with terminal and loading icon
+
 ## [0.1.4] - 2019-08-03
 
 ### Misc